Search for the target method solely by the method name without configuring parameters. Support finding methods by annotating method aliases, so that methods can be found even after confusion. Optimize the object size of annotation classes. Support setting how to backup methods. Basic numerical types can also be fuzzy matched. Add code for a class hook with multiple methods, but not fully implemented.

+++ annotation/src/main/java/qing/albatross/annotation/ConstructorHook.java
@@ -30,7 +30,8 @@ public @interface ConstructorHook {
   Class<?> targetClass() default TargetClass.class;
   int hookerExec() default DEFAULT_OPTION;
   boolean required() default false;
-  int option() default DefOption.DEFAULT;
-  int minSdk() default 0;
-  int maxSdk() default 0;
+  byte option() default DefOption.DEFAULT;
+  byte minSdk() default 0;
+  byte maxSdk() default 0;
 }
+
